Pass const std::string& instead of std::string. Change size_t to GLuint where appropriate.

+++ projects/ROGLL/src/vertex_array.cpp
@@ -30,7 +30,7 @@ namespace ROGLL
 
 		auto attributesVector = attributes.GetAttributes();
 
-		for (size_t i = 0; i < attributesVector.size(); i++)
+		for (GLuint i = 0; i < attributesVector.size(); i++)
 		{
 			auto attribute = attributesVector[i];
